add files for packaging; draft for new result calculation.
