+++ qstnrs/MANIFEST.in
@@ -1,0 +1,3 @@
+include README.rst
+recursive-include questionnaire/static *
+recursive-include questionnaire/templates *
