+++ qstnrs/questionnaire/static/questionnaire/css/style.css
@@ -1,0 +1,3 @@
+li.answer {
+    color: green;
+}
