--- qstnrs/questionnaire/static/questionnaires/style.css
@@ -1,13 +1,0 @@
-ul.errorlist {
-    margin: 0;
-    padding: 0;
-}
-
-.errorlist li {
-    background-color: red;
-    color: white;
-    display: block;
-    font-size: 10px;
-    margin: 0 0 3px;
-    padding: 4px 5px;
-}
