+++ qstnrs/questionnaire/templates/base.html
@@ -1,8 +1,9 @@
+{% load staticfiles %}
 <!DOCTYPE html>
 <html lang="en">
 <meta charset="UTF-8">
 <head>
-    <link rel="stylesheet" type="text/css" href="style.css">
+    <link rel="stylesheet" type="text/css" href="{% static 'questionnaire/css/style.css' %}">
     <title>{% block title %}Questionnaire{% endblock %}</title>
 </head>
 <body>
