+++ qstnrs/questionnaire/templates/page.html
@@ -3,8 +3,8 @@
 {% block title %}Questionnaire{% endblock %}
 
 {% block content %}
-<form action="" method="POST">
 
+{% if questions_list %}
 <form action="" method="post">
     <table>
         {{ form.as_table }}
@@ -21,5 +21,8 @@
     <input type="submit" name="viewResults" value="View Results" />
 {% endif %}
 </form>
+{% else %}
+No questions available.
+{% endif %}
 
 {% endblock %}
