+++ qstnrs/questionnaire/templates/result.html
@@ -3,7 +3,7 @@
 {% block title %}Questionnaire Result{% endblock %}
 
 {% block extrainfo %}
-Your result: {{ score }} <br>
+Your score: {{ score }} <br>
 Your score puts you in the following categories: <br>
 <ul>
     {% for category in categories %}
@@ -13,31 +13,39 @@ Your score puts you in the following categories: <br>
 {% endblock %}
 
 {% block content %}
-{% if different_score > score %}
-For a better score you should have selected: <br>
-<ul>
-    {% for choice in different_choices %}
-        {% for question_text, answer_list in choice.items %}
-            Question: {{ question_text }}
-            {% for answer in answer_list %}
-                Answer: {{ answer }}
+{% if different_choices %}
+    {% if different_score > score %}
+        For a better score you should have selected: </br>
+        {% for choice in different_choices %}
+            {% for question_text, answer_list in choice.items %}
+                    Question: {{ question_text }}
+                    <ul>
+                    {% for answer in answer_list %}
+                        <li class="answer">
+                        Answer: {{ answer }}
+                        </li>
+                    {% endfor %}
+                    </ul>
             {% endfor %}
         {% endfor %}
-    {% endfor %}
-</ul>
-{% elif different_score < score %}
-For a worse score you should have selected: <br>
-<ul>
-    {% for choice in different_choices %}
-        {% for question_text, answer_list in choice.items %}
-            Question: {{ question_text }}
-            {% for answer in answer_list %}
-                Answer: {{ answer }}
+    {% elif different_score < score %}
+        For a worse score you should have selected: <br>
+        {% for choice in different_choices %}
+            {% for question_text, answer_list in choice.items %}
+                    Question: {{ question_text }}
+                     <ul>
+                    {% for answer in answer_list %}
+                        <li class="answer">
+                        Answer: {{ answer }}
+                        </li>
+                    {% endfor %}
+                    </ul>
             {% endfor %}
         {% endfor %}
-    {% endfor %}
-</ul>
+    {% else %}
+        You could not have done better or worse.
+    {% endif %}
 {% else %}
-You could not have done better or worse.
+    No extra selections possible.
 {% endif %}
 {% endblock %}
