+++ qstnrs/questionnaire/urls.py
@@ -3,11 +3,6 @@ from django.contrib import admin
 
 admin.autodiscover()
 
-# url(r'^(?P<questionnaire_id>\d+)/$', 'pages'),
-# url(r'^(?P<questionnaire_id>\d+)/(?P<page_id>\d+)/(?P<question_id>\d+)/$',
-#    'answers')
-# url(r'^(?P<questionnaire_id>\d+)/submit/$', 'submit_page'))
-
 urlpatterns = patterns(
     'questionnaire.views',
     url(r'^$', 'index'),
