+++ qstnrs/setup.py
@@ -1,0 +1,10 @@
+from setuptools import setup
+
+setup(
+      name='Questionnaire',
+      version='0.1',
+      packages=['qstnrs'],
+      include_package_data=True,
+      description='Simple questionnaire app.',
+      author='Bogdan'
+)
